+++ cps/static/js/caliBlur.js
@@ -747,8 +747,8 @@ $("#nav_new a:contains('Recently')").contents().filter(function () {
 shelfText = $(".shelf .discover h2:first").text().replace(":", " —").replace(/\'/g, "");
 $(".shelf .discover h2:first").text(shelfText);
 
-shelfText = $(".shelforder .col-sm-10 .col-sm-6.col-lg-6.col-xs-6 h2:first").text().replace(':', ' —').replace(/\'/g, "");
-$(".shelforder .col-sm-10 .col-sm-6.col-lg-6.col-xs-6 h2:first").text(shelfText);
+shelfText = $(".shelfreorder .col-sm-10 .col-sm-6.col-lg-6.col-xs-6 h2:first").text().replace(':', ' —').replace(/\'/g, "");
+$(".shelfreorder .col-sm-10 .col-sm-6.col-lg-6.col-xs-6 h2:first").text(shelfText);
 
 
 function mobileSupport() {